:truck: api: Move flash files from lib/ to src/
