+++ api/lib/domain/usecases/index.js
@@ -91,7 +91,7 @@ import * as divisionRepository from '../../infrastructure/repositories/division-
 import * as encryptionService from '../../domain/services/encryption-service.js';
 import * as feedbackRepository from '../../infrastructure/repositories/feedback-repository.js';
 import * as finalizedSessionRepository from '../../infrastructure/repositories/sessions/finalized-session-repository.js';
-import * as flashAlgorithmService from '../../domain/services/algorithm-methods/flash.js';
+import * as flashAlgorithmService from '../../../src/certification/flash-certification/domain/services/algorithm-methods/flash.js';
 import * as flashAssessmentResultRepository from '../../infrastructure/repositories/flash-assessment-result-repository.js';
 import * as frameworkRepository from '../../infrastructure/repositories/framework-repository.js';
 import * as groupRepository from '../../infrastructure/repositories/group-repository.js';
