+++ api/src/certification/flash-certification/domain/services/algorithm-methods/flash.js
@@ -172,7 +172,7 @@ function _getInferredChallenges({ challenges, estimatedLevel }) {
  * the challenge with the lowest minimum capability,
  * prioritizing validated challenges over archived ones.
  *
- * @param {import('../../models/Challenge')[]} challenges
+ * @param {import('../../../../../../lib/domain/models/Challenge.js')[]} challenges
  * @returns A list of challenges for scoring inferrence
  */
 function _findChallengesForInferrence(challenges) {
