+++ api/tests/certification/flash-certification/unit/domain/services/algorithm-methods/flash_test.js
@@ -1,6 +1,6 @@
-import { expect, domainBuilder } from '../../../../test-helper.js';
-import * as flash from '../../../../../lib/domain/services/algorithm-methods/flash.js';
-import { AnswerStatus } from '../../../../../lib/domain/models/AnswerStatus.js';
+import { expect, domainBuilder } from '../../../../../../test-helper.js';
+import * as flash from '../../../../../../../src/certification/flash-certification/domain/services/algorithm-methods/flash.js';
+import { AnswerStatus } from '../../../../../../../lib/domain/models/AnswerStatus.js';
 
 describe('Integration | Domain | Algorithm-methods | Flash', function () {
   describe('#getPossibleNextChallenge', function () {
